money converted from int to double

+++ src/com/twitstreet/db/data/StockInPortfolio.java
@@ -3,10 +3,10 @@ package com.twitstreet.db.data;
 public class StockInPortfolio {
 	long stockId;
 	String stockName;
-	int amount;
+	double amount;
 	String pictureUrl;
 	
-	public StockInPortfolio(long stockId, String stockName, int amount, String pictureUrl){
+	public StockInPortfolio(long stockId, String stockName, double amount, String pictureUrl){
 		this.stockId = stockId;
 		this.stockName = stockName;
 		this.amount = amount;
@@ -25,10 +25,10 @@ public class StockInPortfolio {
 	public void setStockName(String stockName) {
 		this.stockName = stockName;
 	}
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
